feat(ui): news dropdown geändert zu table

+++ frontend/src/app/news/news.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, ViewChild, ElementRef, inject } from '@angular/core';
+import { AfterViewInit, Component, OnInit, inject, ViewChild, ElementRef } from '@angular/core';
 import { FormControl, FormGroup, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { INews } from 'src/app/_interface/news';
 import { GlobalDataService } from 'src/app/_service/global-data.service';
@@ -13,6 +13,10 @@ import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { Router } from '@angular/router';
 import { HeaderComponent } from '../_template/header/header.component';
 import { FormatService } from '../helpers/format.service';
+import { MatTableDataSource, MatTableModule } from '@angular/material/table';
+import { MatSortModule } from '@angular/material/sort';
+import { MatPaginator, MatPaginatorModule } from '@angular/material/paginator';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
   selector: 'app-news',
@@ -27,6 +31,10 @@ import { FormatService } from '../helpers/format.service';
     MatOption,
     MatButton,
     MatInputModule,
+    MatTableModule,
+    MatSortModule,
+    MatPaginatorModule,
+    MatIcon,
     MatError,
     NgStyle,
     MatAutocompleteModule
@@ -36,6 +44,7 @@ import { FormatService } from '../helpers/format.service';
 })
 export class NewsComponent implements OnInit {
   @ViewChild('fotoUpload', { static: false }) fotoRef!: ElementRef<HTMLInputElement>;
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
 
   globalDataService = inject(GlobalDataService);
   formatService = inject(FormatService);
@@ -51,6 +60,9 @@ export class NewsComponent implements OnInit {
   filePfad = '';
   fileFound = false;
   btnUploadStatus = false;
+  dataSource = new MatTableDataSource<INews>(this.newsArray);
+  sichtbareSpalten: string[] = ['created_at', 'title', 'typ', 'actions'];
+  pageOptions: any[] = [10, 50, 100]
 
   formAuswahl = new FormGroup({
     news: new FormControl<number | ''>('')
@@ -65,6 +77,10 @@ export class NewsComponent implements OnInit {
     foto_url: new FormControl<string>(''),
   });
 
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
+  }
+
   ngOnInit(): void {
     sessionStorage.setItem('PageNumber', '2');
     sessionStorage.setItem('Page2', 'NEWS');
@@ -76,6 +92,7 @@ export class NewsComponent implements OnInit {
       next: (erg: any) => {
         try {
           this.newsArray = this.convertNewsDate(erg) as INews[];
+          this.dataSource.data = this.newsArray;
         } catch (e: any) {
           this.globalDataService.erstelleMessage('error', e);
         }
@@ -114,10 +131,18 @@ export class NewsComponent implements OnInit {
 
   datenLoeschen(): void {
     const id = this.formModul.controls['id'].value!;
+    this.auswahlLoeschen(id);
+  }
+
+  auswahlLoeschen(id: any): void {
+    if (id === 0) {
+      return;
+    }
     this.globalDataService.delete(this.modul, id).subscribe({
       next: (erg: any) => {
         try {
           this.newsArray = this.newsArray.filter(n => n.id !== id);
+          this.dataSource.data = this.newsArray;
           this.resetFormNachAktion();
           this.globalDataService.erstelleMessage('success', 'News erfolgreich gelöscht!');
         } catch (e: any) {
@@ -128,11 +153,11 @@ export class NewsComponent implements OnInit {
     });
   }
 
-  auswahlBearbeiten(): void {
-    const id = this.formAuswahl.controls['news'].value;
-    if (!id) return;
-
-    const abfrageUrl = `${this.modul}/${id}`;
+  auswahlBearbeiten(element: any): void {
+    if (element.id === 0) {
+      return;
+    }
+    const abfrageUrl = `${this.modul}/${element.id}`;
     this.globalDataService.get(abfrageUrl).subscribe({
       next: (erg: any) => {
         try {
@@ -211,6 +236,7 @@ export class NewsComponent implements OnInit {
           next: (erg: any) => {
             try {
               this.newsArray.push(erg);
+              this.dataSource.data = this.newsArray;
               this.resetFormNachAktion();
               this.globalDataService.erstelleMessage('success', 'News erfolgreich gespeichert!');
             } catch (e: any) {
@@ -225,6 +251,7 @@ export class NewsComponent implements OnInit {
           next: (erg: any) => {
             try {
               this.newsArray.push(erg);
+              this.dataSource.data = this.newsArray;
               this.resetFormNachAktion();
               this.globalDataService.erstelleMessage('success', 'News erfolgreich gespeichert!');
             } catch (e: any) {
@@ -247,6 +274,7 @@ export class NewsComponent implements OnInit {
           next: (erg: any) => {
             try {
               this.newsArray = this.newsArray.map(n => (n.id === erg.id ? erg : n));
+              this.dataSource.data = this.newsArray;
               this.resetFormNachAktion();
               this.globalDataService.erstelleMessage('success', 'News erfolgreich geändert!');
             } catch (e: any) {
@@ -261,6 +289,7 @@ export class NewsComponent implements OnInit {
           next: (erg: any) => {
             try {
               this.newsArray = this.newsArray.map(n => (n.id === erg.id ? erg : n));
+              this.dataSource.data = this.newsArray;
               this.resetFormNachAktion();
               this.globalDataService.erstelleMessage('success', 'News erfolgreich geändert!');
             } catch (e: any) {
